Add unittest for get user profile endpoint, and create new directory for sprint 3 unittest

+++ back-end/server/tests/testUserSignUp.py
@@ -13,12 +13,16 @@ class TestUserSignUp(unittest.TestCase):
         email = "admin@admin.com"
         fn = "Darren"
         ln = "Matthew"
+        brd = "2000/1/1"
+        hometown = "buffalo"
         password_hash = authenticate.generate_hash(uname,pwd)
 
-        #delete user if already exists
-        neo4j.delete_user(uname, password_hash)
+        user = neo4j.get_user(uname)
+        if user == uname:
+            #delete user if already exists
+            neo4j.delete_user(uname, password_hash)
 
-        user = User(uname, fn, ln,email, password_hash)
+        user = User(uname, fn, ln,email, password_hash,brd,hometown)
         # test adding user
         self.assertTrue(neo4j.add_user(user))
         #delete user
